switched to repository pattern. added dist folder to git ignore file
